+++ src/main/java/UI/UIElement.java
@@ -3,10 +3,10 @@ package UI;
 import java.awt.*;
 
 public abstract class UIElement {
-    public int x;
-    public int y;
-    public int width;
-    public int height;
+    public float x;
+    public float y;
+    public float width;
+    public float height;
     public Color bgColor;
 
     public UIElement(int x, int y, int width, int height, Color bgColor) {
